requirements/base.txt
Django
dj-database-url==0.4.1
dj-static==0.0.6
django-appconf==1.0.2
django-cors-middleware==1.3.1
django-import-export==0.5.1
django-s3direct==0.4.4
django-storages==1.5.1
django-toolbelt==0.0.1
djangorestframework==3.4.6
gunicorn==19.6.0
Pillow==3.4.2
plivo==0.10.3
psycopg2==2.6.2
pytz==2016.7
requests==2.12.3
twilio==4.6.0
tinify==1.5.0
boto3==1.4.2
dj-database-url==0.4.1
dj-static==0.0.6
django-redis-cache==1.7.1
django-rq==0.9.4
django-s3direct==0.4.2
django-storages==1.5.1
django-taggit==0.21.2
drf-extensions==0.3.1
simplejson==3.8.2




